Temporarily disable github workflows
Currently, we don't really support using nix-thunk except by installing
it via nix. In order to support non-nix installs, we will have to make
the build depend on nix, so that it can create a store path for its
vendored nixpkgs checkout, and keep a GC root around somewhere.

The number of users of nix-thunk who are not installing it through nix
might number in the high zeroes.